requirements.txt
python-telegram-bot==13.15
requests==2.32.3
environs==11.0.0
